Attempt to rewrite things
This unifies the approach between different commands, and makes them
actual subcommands instead of flags.

All components (or subcommands) in this setup share same flags and same
execution logic, thus they were unified. This brings the downside of
requiring another rewrite in case we have a subdommand that uses
different flow, but it is ok for now.